Extend list of hacks for reading non-standard complied 1d WCS fits files
In the current code, there is a hack to allow "LINEAR" as a spectral
axis. That's clearly not complient with the FITS WCS standard as defined in
https://www.aanda.org/articles/aa/full/2006/05/aa3818-05/aa3818-05.html
and might just be motivated by the GMOS spectrum currently used as
example for this loader.
This code change makes ths list of keywords that will be treated as
spectral extensible and adds "Wavelength" (which is much closer to
"WAVE" as required in the standard than "LINEAR").