Explain how to make a continuum normalized spectrum
In the fitting section, the docs talk quite a bit about how continuum
normalized spectra are useful and how the tools help to make them.
However, it is not shown how to actually normalize the spectrum.
I opened #415 because I could not find the answer after experimenting.
This doc change adds a simple example.