Add helper functions to control spectral shifting
The setters for Spectrum1d.redshift and Spectrum1d.radial_velocity shift
the spectral axis, which from #820 is not always what users expect. This
adds three helpers, one which does the current shifting (and is
explicitly documented to do that), and two which simply change the
values of redshift and radial_velocity.