For ``tabular_fits`` output, force ``uncertainty`` to StdDev
In the ``tabular_fits_writer``, the uncertainty is written out as the standard
deviation.  Some readers, however, load 1D spectra into ``Spectrum1D`` with the
uncertainty specified as the inverse variance.  Without explicitly converting
the uncertainty to standard deviation, the code raises a
``astropy.units.core.UnitConversionError`` for these cases when trying to
conform the units of the uncertainty to those of the spectral flux array.

This commit utilizes the ``represent_as`` functionality of
``astropy.nddata.NDUncertainty`` arrays to force the representation of the
uncertainty in standard deviation form before checking units and writing to
FITS.

	modified:   specutils/io/default_loaders/tabular_fits.py