Minor changes to DC loaders for future surveys (#1122)
* Include all remote data in coverage run

This should reduce the number of tests skipped, and resolve some of the
weirdness with coverage seen in #1122.

* Minor changes to DC loaders for future surveys

This makes it easier to use a fallback header and access the last
spectra from a wrapper loader. These changes are to support loading some
of the IFS surveys (such as SAMI) that Data Central hosts.

* Add SAMI loaders for testing updated DC loaders

This adds an initial version of the future SAMI loaders, in order to
test the changes to the lower level DC loaders.