Add linear and loglinear wavelength solutions to (IRAF) fits-wcs
This adds linear and log-linear solutions to IRAF MULTISPEC format. Since the implementation is shares with the simpler IRAF formats (which had only linear, but not log-linear before), they gain the log-linear solutions as well.

This is one step towards #229. The MUTLISPEC format is often used for echelle spectra and linear solutions for that are common in the wild (e.g. KECK HIRES pipeline).